Separated CLI and Main logic now people can import csv2docx as library
