+++ pyproject.toml
@@ -16,7 +16,7 @@ click = "^7.1.2"
 [tool.poetry.dev-dependencies]
 
 [tool.poetry.scripts]
-convert = "csv2docx.cli:convert"
+convert = "csv2docx.cli:main"
 
 [build-system]
 requires = ["poetry>=0.12"]
